# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * VGC Teambuilding Helper API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface PokemonSummary {
  name: string;
  dexNumber: number;
  types: string[];
  spriteUrl: string;
  isLegendary?: boolean;
}

export interface BaseStats {
  hp: number;
  attack: number;
  defense: number;
  specialAttack: number;
  specialDefense: number;
  speed: number;
}

export interface PokemonDetail {
  name: string;
  dexNumber: number;
  types: string[];
  baseStats: BaseStats;
  abilities: string[];
  spriteUrl: string;
  artworkUrl: string;
  learnableMoves?: string[];
  weightKg?: number;
}

export interface Regulation {
  id: string;
  name: string;
  label: string;
  description: string;
  isActive: boolean;
  startDate: string;
  /** @nullable */
  endDate?: string | null;
  allowedSeriesName?: string;
  restrictedLegendariesAllowed?: number;
  notes?: string;
}

export interface Archetype {
  id: string;
  name: string;
  description: string;
  category: string;
  coreStrategy: string;
  keyPokemon: string[];
  strengths: string[];
  weaknesses: string[];
  counterplay?: string;
  dangerLevel?: number;
}

export interface StatSpread {
  hp?: number;
  attack?: number;
  defense?: number;
  specialAttack?: number;
  specialDefense?: number;
  speed?: number;
}

export interface TeamSlot {
  slot: number;
  pokemonName: string;
  /** @maxItems 4 */
  moves: string[];
  ability: string;
  item: string;
  nature: string;
  teraType: string;
  evs?: StatSpread;
  ivs?: StatSpread;
}

export interface Team {
  id: number;
  name: string;
  regulation: string;
  /** @nullable */
  description?: string | null;
  slots: TeamSlot[];
  createdAt: string;
  updatedAt: string;
}

export interface TeamInput {
  name: string;
  regulation: string;
  description?: string;
  slots: TeamSlot[];
}

export interface TypeDefenseEntry {
  type: string;
  multiplier: number;
  pokemonResisting: string[];
}

export interface SpeedTierEntry {
  pokemonName: string;
  baseSpeed: number;
  effectiveSpeed: number;
  nature?: string;
}

export interface WeaknessSummaryEntry {
  type: string;
  count: number;
  vulnerablePokemon: string[];
}

export interface TeamAnalysis {
  teamId: number;
  typeDefenses: TypeDefenseEntry[];
  offensiveCoverage: string[];
  speedTiers: SpeedTierEntry[];
  weaknessSummary: WeaknessSummaryEntry[];
  suggestions: string[];
  terrainControl?: boolean;
  weatherControl?: boolean;
  hasTrickRoom?: boolean;
  redirectSupport?: boolean;
}

export interface MatchupResult {
  archetypeId: string;
  archetypeName: string;
  rating: number;
  verdict: string;
  advantages: string[];
  disadvantages: string[];
  keyThreats: string[];
  winConditions: string[];
}

export type SearchPokemonParams = {
q: string;
regulation?: string;
};