# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: VGC Teambuilding Helper API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /pokemon/search:
    get:
      operationId: searchPokemon
      tags: [pokemon]
      summary: Search Pokemon by name
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
        - name: regulation
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of matching Pokemon
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PokemonSummary"
  /pokemon/{name}:
    get:
      operationId: getPokemon
      tags: [pokemon]
      summary: Get full Pokemon data
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Pokemon detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PokemonDetail"
        "404":
          description: Not found
...
  /teams/{id}/analyze:
    post:
      operationId: analyzeTeam
      tags: [analysis]
      summary: Analyze a team's coverage, weaknesses, and speed tiers
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Analysis results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeamAnalysis"
  /teams/{id}/simulate:
    post:
      operationId: simulateTeam
      tags: [analysis]
      summary: Simulate team against all meta archetypes
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Simulation results
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MatchupResult"
components:
  schemas:
...
    PokemonSummary:
      type: object
      required: [name, dexNumber, types, spriteUrl]
      properties:
        name:
          type: string
        dexNumber:
          type: integer
        types:
          type: array
          items:
            type: string
        spriteUrl:
          type: string
        isLegendary:
          type: boolean
    PokemonDetail:
      type: object
      required: [name, dexNumber, types, baseStats, abilities, spriteUrl, artworkUrl]
      properties:
        name:
          type: string
        dexNumber:
          type: integer
        types:
          type: array
          items:
            type: string
        baseStats:
          $ref: "#/components/schemas/BaseStats"
        abilities:
          type: array
          items:
            type: string
        spriteUrl:
          type: string
        artworkUrl:
          type: string
        learnableMoves:
          type: array
          items:
            type: string
        weightKg:
          type: number
    BaseStats:
      type: object
      required: [hp, attack, defense, specialAttack, specialDefense, speed]
      properties:
        hp:
          type: integer
        attack:
          type: integer
        defense:
          type: integer
        specialAttack:
          type: integer
        specialDefense:
          type: integer
        speed:
          type: integer
...
    TeamAnalysis:
      type: object
      required: [teamId, typeDefenses, offensiveCoverage, speedTiers, weaknessSummary, suggestions]
      properties:
        teamId:
          type: integer
        typeDefenses:
          type: array
          items:
            $ref: "#/components/schemas/TypeDefenseEntry"
        offensiveCoverage:
          type: array
          items:
            type: string
        speedTiers:
          type: array
          items:
            $ref: "#/components/schemas/SpeedTierEntry"
        weaknessSummary:
          type: array
          items:
            $ref: "#/components/schemas/WeaknessSummaryEntry"
        suggestions:
          type: array
          items:
            type: string
        terrainControl:
          type: boolean
        weatherControl:
          type: boolean
        hasTrickRoom:
          type: boolean
        redirectSupport:
          type: boolean
    TypeDefenseEntry:
      type: object
      required: [type, multiplier, pokemonResisting]
      properties:
        type:
          type: string
        multiplier:
          type: number
        pokemonResisting:
          type: array
          items:
            type: string
    SpeedTierEntry:
      type: object
      required: [pokemonName, baseSpeed, effectiveSpeed]
      properties:
        pokemonName:
          type: string
        baseSpeed:
          type: integer
        effectiveSpeed:
          type: integer
        nature:
          type: string
    WeaknessSummaryEntry:
      type: object
      required: [type, count, vulnerablePokemon]
      properties:
        type:
          type: string
        count:
          type: integer
        vulnerablePokemon:
          type: array
          items:
            type: string
    MatchupResult:
      type: object
      required: [archetypeId, archetypeName, rating, verdict, advantages, disadvantages, keyThreats, winConditions]
      properties:
        archetypeId:
          type: string
        archetypeName:
          type: string
        rating:
          type: integer
        verdict:
          type: string
        advantages:
          type: array
          items:
            type: string
        disadvantages:
          type: array
          items:
            type: string
        keyThreats:
          type: array
          items:
            type: string
        winConditions:
          type: array
          items:
            type: string