# lib/api-zod/src/generated/types/baseStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * VGC Teambuilding Helper API
 * OpenAPI spec version: 0.1.0
 */

export interface BaseStats {
  hp: number;
  attack: number;
  defense: number;
  specialAttack: number;
  specialDefense: number;
  speed: number;
}