# lib/api-zod/src/generated/types/matchupResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * VGC Teambuilding Helper API
 * OpenAPI spec version: 0.1.0
 */

export interface MatchupResult {
  archetypeId: string;
  archetypeName: string;
  rating: number;
  verdict: string;
  advantages: string[];
  disadvantages: string[];
  keyThreats: string[];
  winConditions: string[];
}