# lib/api-zod/src/generated/types/pokemonDetail.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * VGC Teambuilding Helper API
 * OpenAPI spec version: 0.1.0
 */
import type { BaseStats } from './baseStats';

export interface PokemonDetail {
  name: string;
  dexNumber: number;
  types: string[];
  baseStats: BaseStats;
  abilities: string[];
  spriteUrl: string;
  artworkUrl: string;
  learnableMoves?: string[];
  weightKg?: number;
}