# lib/api-zod/src/generated/types/pokemonSummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * VGC Teambuilding Helper API
 * OpenAPI spec version: 0.1.0
 */

export interface PokemonSummary {
  name: string;
  dexNumber: number;
  types: string[];
  spriteUrl: string;
  isLegendary?: boolean;
}