# lib/api-zod/src/generated/types/speedTierEntry.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * VGC Teambuilding Helper API
 * OpenAPI spec version: 0.1.0
 */

export interface SpeedTierEntry {
  pokemonName: string;
  baseSpeed: number;
  effectiveSpeed: number;
  nature?: string;
}