# lib/api-zod/src/generated/types/weaknessSummaryEntry.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * VGC Teambuilding Helper API
 * OpenAPI spec version: 0.1.0
 */

export interface WeaknessSummaryEntry {
  type: string;
  count: number;
  vulnerablePokemon: string[];
}